perf(cli): enable code splitting and deferred UI loading
